refactor modules folder structure
