+++ src/homepage/homepage.css
@@ -8,7 +8,7 @@ body {
   font-family: Arial, Helvetica, sans-serif;
   background-image:
     linear-gradient(to bottom, rgba(0, 0, 0, 0.8), rgba(5, 5, 5, 0.6)),
-    url('./homepage-background.jpg');
+    url('../assets/homepage-background.jpg');
   background-repeat: no-repeat;
   background-size: cover;
   height: 100vh;
